+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.6.0] - 2020-02-10
+### Added
+- CLI `--help` and related arguments.
+- `--move` option. Default is copy, but this CLI option activates copy instead.
+
 ### Fixed
 - Added missing docstrings.